Improved code structure
Now shows paste status when pasting from explorer view (folders tab)

+++ src/commands/paste.ts
@@ -25,10 +25,10 @@ export default async function paste(dir: string | undefined) {
 
         if(connection?.label){
             if(dir === undefined){
-                const clipboard = await getClipboardContent(config, connection.label);
                 const editor = window.activeTextEditor;
                 if(!editor) return window.showErrorMessage("No active editor found.");
-
+                
+                const clipboard = await getClipboardContent(config, connection.label);
                 if(clipboard.startsWith("vscode")){
                     const folderName = await window.showInputBox({ prompt: "Save To Folder" });
                     if(!folderName) return window.showWarningMessage(`Paste cancelled.`);
@@ -36,14 +36,7 @@ export default async function paste(dir: string | undefined) {
                     const saveDir = workspace.getWorkspaceFolder(editor.document.uri);
                     if(!saveDir) return window.showWarningMessage('Paste error.');
 
-                    const directories = clipboard.split("/****************").filter(t => t !== "/****************").filter(Boolean);
-                    await workspace.fs.createDirectory(Uri.file(path.join(saveDir.uri.path, folderName)));
-                    for(let i = 1; i < directories.length; i++){
-                        if(directories[i].startsWith("\n//* vscode ")){
-                            const filePath = Uri.joinPath(Uri.file(saveDir.uri.path), folderName, directories[i].replace("//* vscode ", '').trim());
-                            await workspace.fs.writeFile(filePath, Buffer.from(directories[i+1].trim(), 'utf-8'));
-                        }
-                    }
+                    vscodeClipboard(saveDir.uri.path, folderName, clipboard, connection.label);
                 }else{
                     const selection = editor.selection;
                     const pasted = await editor.edit(editBuilder => {
@@ -55,17 +48,25 @@ export default async function paste(dir: string | undefined) {
                     window.showInformationMessage(`Pasted ${connection.label} at line ${selection.active.line + 1}`);
                 }
             }else{
-                const fileName = await window.showInputBox({ prompt: "Save As" });
-                if(!fileName) return window.showWarningMessage(`Paste cancelled.`);
-
                 const clipboard = await getClipboardContent(config, connection.label);
-                
-                const filePath = Uri.file(path.join(dir, fileName));
-                await workspace.fs.writeFile(filePath, Buffer.from(clipboard, 'utf-8'));
+                if(clipboard.startsWith("vscode")){
+                    const folderName = await window.showInputBox({ prompt: "Save To Folder" });
+                    if(!folderName) return window.showWarningMessage(`Paste cancelled.`);
 
-                window.showInformationMessage(`Pasted ${connection.label} at ${fileName}`);
-                const createdFile = await workspace.openTextDocument(filePath);
-                await window.showTextDocument(createdFile);
+                    vscodeClipboard(dir, folderName, clipboard, connection.label);
+                }else{
+                    const fileName = await window.showInputBox({ prompt: "Save As" });
+                    if(!fileName) return window.showWarningMessage(`Paste cancelled.`);
+
+                    const clipboard = await getClipboardContent(config, connection.label);
+                    
+                    const filePath = Uri.file(path.join(dir, fileName));
+                    await workspace.fs.writeFile(filePath, Buffer.from(clipboard, 'utf-8'));
+
+                    window.showInformationMessage(`Pasted ${connection.label} at ${fileName}`);
+                    const createdFile = await workspace.openTextDocument(filePath);
+                    await window.showTextDocument(createdFile);
+                }
             }
         }else{
             window.showWarningMessage('Paste cancelled.');
@@ -73,4 +74,19 @@ export default async function paste(dir: string | undefined) {
     }catch{
         window.showErrorMessage('Paste error.');
     }
+}
+
+async function vscodeClipboard(saveDir: string, folderName: string, clipboard: string, connection: string) {
+    const directories = clipboard.split("/****************").filter(t => t !== "/****************").filter(Boolean);
+    await workspace.fs.createDirectory(Uri.file(path.join(saveDir, folderName)));
+    for(let i = 1; i < directories.length; i++){
+        if(directories[i].startsWith("\n//* vscode ")){
+            const filePath = Uri.joinPath(Uri.file(saveDir), folderName, directories[i].replace("//* vscode ", '').trim());
+            await workspace.fs.writeFile(filePath, Buffer.from(directories[i+1].trim(), 'utf-8'));
+            const createdFile = await workspace.openTextDocument(filePath);
+            await window.showTextDocument(createdFile);
+        }
+    }
+
+    window.showInformationMessage(`Pasted ${connection} at ${folderName}`);
 }
